refactor : CommentController, LikeController 디렉토리 수
